third_party/blink/web_tests/virtual/select-parser-relaxation-opt-out/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-optgroup-arrow-keys.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Keyboard navigation forwards and backwards should visit each option with optgroups.
  assert_equals: appearance:base-select must be supported to run this test. expected "base-select" but got "auto"
Harness: the test ran to completion.
